==> notebooks/01_load_and_validate_raw_data.ipynb <==
from pathlib import Path
import requests
import zipfile
import pandas as pd
import pyarrow.parquet as pq

def download_one_file_of_raw_data(year: int) -> Path:
    """"""
    URL = f'https://cdn.buenosaires.gob.ar/datosabiertos/datasets/transporte-y-obras-publicas/bicicletas-publicas/recorridos-realizados-{year}.zip'
    response= requests.get(URL) #, stream=True)

    
    if response.status_code == 200:
        # with open(nombre_archivo_zip, 'wb') as file:
        #     for chunk in response.iter_content(chunk_size=8192):  # Tamaño del búfer ajustado a 8192 bytes
        #         file.write(chunk)
        path = f'../data/raw/recorridos-realizados-{year}.zip'
        open(path, "wb").write(response.content)     
        print(f'descargado año {year}')

        return path
    else:
        raise Exception(f'{URL} is not available')

import pyarrow as pa

def unzip_and_convert_csv_to_parquet(year: int) -> Path:
    nombre_archivo_zip = f"../data/raw/recorridos-realizados-{year}.zip"
        # Descomprimir el archivo zip
    with zipfile.ZipFile(nombre_archivo_zip, 'r') as archivo_zip:

        # Extraer el archivo CSV del zip
        nombre_archivo_csv = archivo_zip.namelist()[0]  # Suponiendo que el archivo CSV es el primer archivo en el zip
        archivo_zip.extractall(f"../data/raw/")

        # Leer el archivo CSV con pandas
        df = pd.read_csv(f"../data/raw/{nombre_archivo_csv}", delimiter=',', decimal=".")

        # Convertir el DataFrame a formato parquet
        nombre_archivo_parquet = f"rides_{year}.parquet"
        table = pa.Table.from_pandas(df)
        pq.write_table(table, f"../data/raw/{nombre_archivo_parquet}")



        # table = pq.Table.from_pandas(df)
        # pq.write_table(table, nombre_archivo_parquet)
        #pq.write_table(pq.Table.from_pandas(df), nombre_archivo_parquet)
        path = f'../data/raw/rides_{year}.parquet'
    return path
    

# Conversion manual de archivos descargados, ha sido un apaño para resolverlo !!!!!!  importante

# nombre_archivo_zip = f"../data/raw/recorridos-realizados-2025.zip"

#  # Leer el archivo CSV con pandas
# df = pd.read_csv(f"../data/raw/trips_2025.csv", delimiter=',', decimal=".")

# # Convertir el DataFrame a formato parquet
# nombre_archivo_parquet = f"rides_2025.parquet"
# table = pa.Table.from_pandas(df)
# pq.write_table(table, f"../data/raw/{nombre_archivo_parquet}")


# path = f'../data/raw/rides_2025.parquet'

download_one_file_of_raw_data(year=2022)

unzip_and_convert_csv_to_parquet(year=2022)

import pandas as pd

rides = pd.read_parquet('../data/raw/rides_2024.parquet')

rides.head(10)

import pandas as pd

rides = pd.read_parquet('../data/raw/rides_2022.parquet')

rides.head(10)

rides.shape

rides.id_estacion_origen.nunique()

#rides[['fecha_origen_recorrido', 'id_estacion_origen']].info()

#Nos quedamos sólo con las columnas que nos interesan y las renombramos
rides = rides[['fecha_origen_recorrido', 'id_estacion_origen']]
# Eliminar la parte "BAEcobici" y convertir a tipo int
rides['id_estacion_origen'] = rides['id_estacion_origen'].str.replace('BAEcobici', '').astype(int)

rides.rename(columns={
    'fecha_origen_recorrido': 'pickup_datetime',
    'id_estacion_origen': 'pickup_location_id',
}, inplace=True)

rides['pickup_datetime'] = pd.to_datetime(rides['pickup_datetime'],format='%Y-%m-%d %H:%M:%S')
rides.head(10)

rides.pickup_location_id.nunique()

rides['pickup_datetime'].describe()

rides = rides[rides.pickup_datetime >= '2024-01-01']
rides = rides[rides.pickup_datetime < '2025-01-01']
rides['pickup_datetime'].describe()

rides.to_parquet('../data/transformed/validated_rides_2024.parquet')
