==> notebooks/04_transform_raw_data_into_features_and_targets.ipynb <==
%reload_ext autoreload
%autoreload 2

from src.data import load_raw_data

rides = load_raw_data(year=2022)
rides

rides.info()

import pandas as pd
nulos = pd.DataFrame(rides.isnull().sum(), columns=['Nulos'])
nulos.Nulos.unique()

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)
ts_data


ts_data.pickup_location_id.unique().size

import pandas as pd
nulos = pd.DataFrame(ts_data.isnull().sum(), columns=['Nulos'])
nulos.Nulos.unique()

from src.data import transform_ts_data_into_features_and_target

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len=24*28*1, # one month
    step_size=24,
    output_seq_len=36
)

print(f'{features.shape=}')
print(f'{targets.shape=}')

import pandas as pd
nulos = pd.DataFrame(features.isnull().sum(), columns=['Nulos'])
nulos.Nulos.unique()

targets.tail()

import pandas as pd
#tabular_data = features
#tabular_data = targets
tabular_data = pd.concat([features, targets], axis=1)

from src.paths import TRANSFORMED_DATA_DIR
tabular_data.to_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')

tabular_data.head()

import pandas as pd
nulos = pd.DataFrame(tabular_data.isnull().sum(), columns=['Nulos'])
nulos.Nulos.unique()
