==> notebooks/05_visualize_training_data.ipynb <==
%reload_ext autoreload
%autoreload 2

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

features_and_target = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
features_and_target

features = features_and_target.filter(regex='previous|pickup')
targets = features_and_target.filter(like='next')

features_ = features.iloc[500]
target_ = targets.iloc[500]
    
ts_columns_features = [c for c in features.columns if c.startswith('rides_previous_')]# or c.startswith('rides_next_')]
ts_columns_targets = [c for c in targets.columns if c.startswith('rides_next_')]
ts_values_features = [features_[c] for c in ts_columns_features] # + target_.values.tolist()
ts_values_targets = [target_[c] for c in ts_columns_targets] 

len(ts_columns_features)

+672+36

len(ts_values_features)

len(ts_columns_features)

24*7*4

len(ts_values_targets)

features_['pickup_hour']

from datetime import timedelta
pd.date_range(
        features_['pickup_hour'] - timedelta(hours=len(ts_columns_features)),
        features_['pickup_hour'] - timedelta(hours=1),
        freq='H'
    )

from datetime import timedelta
ts_dates_features = pd.date_range(
        features_['pickup_hour'] - timedelta(hours=len(ts_columns_features)),
        features_['pickup_hour'] - timedelta(hours=1),
        freq='H'
    )
ts_dates_targets = pd.date_range(
        features_['pickup_hour'],
        features_['pickup_hour'] + timedelta(hours=len(ts_columns_targets)-1),
        freq='H'
    )

len(ts_dates_features)

len(ts_dates_targets)

import plotly.express as px 
# line plot with past values
title = f'Pick up hour={features_["pickup_hour"]}, location_id={features_["pickup_location_id"]}'
fig = px.line(
        x=ts_dates_features, y=ts_values_features,
        template='plotly_dark',
        markers=True, title=title
    )
# green line for the values we wanna predict
targets_fig = px.line(x=ts_dates_targets, y=target_.values.tolist(),
                template='plotly_dark', 
                markers=True, title='actual values')
targets_fig.update_traces(line_color='green')

fig.add_traces(targets_fig.data)
#fig.show()



from src.plot import plot_one_sample

plot_one_sample(
    features,
    targets,
    example_id=500,
)
