==> notebooks/07_xgboost_model.ipynb <==
import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2022, 6, 1, 0, 0, 0),
    targets_columns_names=[c for c in df.columns if c.startswith('rides_next_')]
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

import xgboost as xgb

# use only past rides data
past_rides_columns = [c for c in X_train.columns if c.startswith('rides_')]
X_train_only_numeric = X_train[past_rides_columns]

model = xgb.XGBRegressor()
model.fit(X_train_only_numeric, y_train)

X_test_only_numeric = X_test[past_rides_columns]
predictions = model.predict(X_test_only_numeric)
predictions

# Como da algun valor negativo como resultado eliminamos las predicciones negativas y los intervalos negativos reemplazando por cero
import numpy as np

#Limitar los valores predichos a un mínimo de cero
predictions = np.clip(predictions, 0, None)

predictions

predictions=pd.DataFrame(predictions, columns=[f'rides_next_{i+1}_hour' for i in range(36)])
predictions

from sklearn.metrics import mean_absolute_error
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

#from sktime.forecasting.compose import MultioutputForecaster
#from sktime.forecasting.model_selection import temporal_train_test_split
#from sktime.forecasting.xgboost import XGBoostRegressor

import numpy as np
from tqdm import tqdm
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import StandardScaler
#from xgboost import XGBRegressor

len(X_test_only_numeric)

X_test_only_numeric.loc[1]

X_test_only_numeric.loc[500]
