==> notebooks/09_lightgbm_model_with_feature_engineering.ipynb <==
import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2022, 6, 1, 0, 0, 0),
    targets_columns_names=[c for c in df.columns if c.startswith('rides_next_')]
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

def average_rides_last_4_weeks(X: pd.DataFrame) -> pd.DataFrame:
    """
    Adds one column with the average rides from
    - 7 days ago
    - 14 days ago
    - 21 days ago
    - 28 days ago
    """
    X['average_rides_last_4_weeks'] = 0.25*(
        X[f'rides_previous_{7*24}_hour'] + \
        X[f'rides_previous_{2*7*24}_hour'] + \
        X[f'rides_previous_{3*7*24}_hour'] + \
        X[f'rides_previous_{4*7*24}_hour']
    )
    return X

from sklearn.preprocessing import FunctionTransformer

add_feature_average_rides_last_4_weeks = FunctionTransformer(
    average_rides_last_4_weeks, validate=False)

add_feature_average_rides_last_4_weeks.fit_transform(X_train)

import pandas as pd
from src.paths import RAW_DATA_DIR

def latitude_and_longitude(X: pd.DataFrame) -> pd.DataFrame:
    """
    Adds two columns with the latitude and longitude from pickup_location_id
    
    """
    raw_data_rides = pd.read_parquet(RAW_DATA_DIR / 'rides_2022.parquet')

    #Nos quedamos sólo con las columnas que nos interesan y las renombramos
    raw_data_rides = raw_data_rides[['id_estacion_origen', 'lat_estacion_origen', 'long_estacion_origen']]
    raw_data_rides['id_estacion_origen'] = raw_data_rides['id_estacion_origen'].str.replace('BAEcobici', '').astype(int)
    raw_data_rides = raw_data_rides.drop_duplicates().reset_index(drop=True)
    raw_data_rides.rename(columns={
    'id_estacion_origen': 'pickup_location_id',
    'lat_estacion_origen': 'latitude',
    'long_estacion_origen': 'longitude'
    }, inplace=True)

    # Combinar la información de latitud y longitud en X
    X = X.merge(raw_data_rides, on='pickup_location_id', how='left')

    # Eliminar la columna 'pickup_location_id'
    #X.drop('pickup_location_id', axis=1, inplace=True)

    return X

from sklearn.preprocessing import FunctionTransformer

add_feature_latitude_and_longitude = FunctionTransformer(
    latitude_and_longitude, validate=False)

add_feature_latitude_and_longitude.fit_transform(X_train)

from sklearn.base import BaseEstimator, TransformerMixin

class TemporalFeaturesEngineer(BaseEstimator, TransformerMixin):

    def fit(self, X, y=None):
        return self
    
    def transform(self, X, y=None):
        
        X_ = X.copy()
        
        # Generate numeric columns from datetime
        X_["hour"] = X_['pickup_hour'].dt.hour
        X_["day_of_week"] = X_['pickup_hour'].dt.dayofweek

                        

        return X_.drop(columns=['pickup_hour'])

add_temporal_features = TemporalFeaturesEngineer()
add_temporal_features.fit_transform(X_train)

import lightgbm as lgb
from sklearn.multioutput import MultiOutputRegressor

from sklearn.pipeline import make_pipeline

pipeline = make_pipeline(
    add_feature_average_rides_last_4_weeks,
    add_feature_latitude_and_longitude,
    add_temporal_features,
    MultiOutputRegressor(lgb.LGBMRegressor(force_col_wise=True))
)
pipeline.fit(X_train, y_train)

X_train

predictions = pipeline.predict(X_test)

from sklearn.metrics import mean_absolute_error
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

# Como da algun valor negativo como resultado eliminamos las predicciones negativas y los intervalos negativos reemplazando por cero
import numpy as np

#Limitar los valores predichos a un mínimo de cero
predictions = np.clip(predictions, 0, None)

from sklearn.metrics import mean_absolute_error
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')
