==> notebooks/10_lightgbm_model_with_hyperparameter_tuning.ipynb <==
%reload_ext autoreload
%autoreload 2

import warnings
warnings.filterwarnings("ignore")

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2022, 6, 1, 0, 0, 0),
    targets_columns_names=[c for c in df.columns if c.startswith('rides_next_')]
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

import numpy as np
from sklearn.model_selection import KFold, TimeSeriesSplit
from sklearn.pipeline import make_pipeline
from sklearn.metrics import mean_absolute_error
import optuna

from src.model import get_pipeline

def objective(trial: optuna.trial.Trial) -> float:
    """
    Given a set of hyper-parameters, it trains a model and computes an average
    validation error based on a TimeSeriesSplit
    """
    # pick hyper-parameters
    hyperparams = {
        "metric": 'mae',
        "verbose": -1,
        "num_leaves": trial.suggest_int("num_leaves", 2, 256),
        "feature_fraction": trial.suggest_float("feature_fraction", 0.2, 1.0),
        "bagging_fraction": trial.suggest_float("bagging_fraction", 0.2, 1.0),
        "min_child_samples": trial.suggest_int("min_child_samples", 3, 100),   
    }
       
    tss = TimeSeriesSplit(n_splits=2)
    scores = []
    for train_index, val_index in tss.split(X_train):

        # split data for training and validation
        X_train_, X_val_ = X_train.iloc[train_index, :], X_train.iloc[val_index,:]
        y_train_, y_val_ = y_train.iloc[train_index], y_train.iloc[val_index]
        
        # train the model
        pipeline = get_pipeline(**hyperparams)
        pipeline.fit(X_train_, y_train_)
        
        # evaluate the model
        y_pred = pipeline.predict(X_val_)
        mae = mean_absolute_error(y_val_, y_pred)

        scores.append(mae)
   
    # Return the mean score
    return np.array(scores).mean()

study = optuna.create_study(direction="minimize")
study.optimize(objective, n_trials=5)

best_params = study.best_trial.params
print(f'{best_params=}')

pipeline = get_pipeline(**best_params)
pipeline.fit(X_train, y_train)

predictions = pipeline.predict(X_test)
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

# Como da algun valor negativo como resultado eliminamos las predicciones negativas y los intervalos negativos reemplazando por cero
import numpy as np

#Limitar los valores predichos a un mínimo de cero
predictions = np.clip(predictions, 0, None)

from sklearn.metrics import mean_absolute_error
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

from src.plot import plot_one_sample

plot_one_sample(
    features=X_test,
    targets=y_test,
    example_id=500,
    predictions=pd.DataFrame(predictions)
)

Se agrega este paso aquí en el que vamos a crear un feature group para almacenar los datos de pickup_location_id, latitud y longitud. Por tanto vamos a modificar la funcion latitud y longitud del script "model" para que solo sea una consulta al feature group


from src.paths import RAW_DATA_DIR
import pandas as pd

raw_data_rides = pd.read_parquet(RAW_DATA_DIR / 'rides_2022.parquet')

#Nos quedamos sólo con las columnas que nos interesan y las renombramos
raw_data_rides = raw_data_rides[['id_estacion_origen', 'lat_estacion_origen', 'long_estacion_origen']]
raw_data_rides['id_estacion_origen'] = raw_data_rides['id_estacion_origen'].str.replace('BAEcobici', '').astype(int)
raw_data_rides = raw_data_rides.drop_duplicates().reset_index(drop=True)
raw_data_rides.rename(columns={
    'id_estacion_origen': 'pickup_location_id',
    'lat_estacion_origen': 'latitude',
    'long_estacion_origen': 'longitude'
    }, inplace=True)



# Combinar la información de latitud y longitud en X
#X = X.merge(raw_data_rides, on='pickup_location_id', how='left')

raw_data_rides.head()

from src.feature_store_api import get_feature_store
import src.config as config

# connect to the feature group
feature_group = get_feature_store().get_or_create_feature_group(
    name=config.FEATURE_GROUP_LAT_LONG,
    version=1,
    description="Latitud y longitud de pickup_location_id",
    primary_key = ['pickup_location_id', 'latitude', 'longitude']
    #event_time='pickup_hour',
)

feature_group.insert(raw_data_rides, write_options={"wait_for_job": False})

import hopsworks

# connect to the project
project = hopsworks.login(
    project=config.HOPSWORKS_PROJECT_NAME,
    api_key_value=config.HOPSWORKS_API_KEY
)

# connect to the feature store
feature_store = project.get_feature_store()

# # connect to the feature group
# feature_group = feature_store.get_feature_group(
#     name=config.FEATURE_GROUP_NAME,
#     version=config.FEATURE_GROUP_VERSION,
# )

# create feature view (if it doesn't exist yet)
# This feature view only uses on feature group, so the query is trivial
try:
    # create feature view if it doesn't exist yet
    feature_store.create_feature_view(
        name=config.FEATURE_VIEW_LAT_LONG,        
        query=feature_group.select_all()
    )
except:
    print('Feature view already existed. Skip creation.')


# get feature view
feature_view = feature_store.get_feature_view(
    name=config.FEATURE_VIEW_LAT_LONG
)

feature_store = get_feature_store()
feature_view = feature_store.get_feature_view(
        name=config.FEATURE_VIEW_LAT_LONG
    )
raw_data_rides= feature_view.get_batch_data()
