==> notebooks/14_inference_pipeline.ipynb <==
%load_ext autoreload
%autoreload 2

from datetime import datetime, timedelta
import pandas as pd

current_date = pd.to_datetime(datetime.utcnow(), utc=True).floor('H') # - timedelta(hours=1)
print(f'{current_date=}')
#current_date = pd.Timestamp('2023-11-10 10:00:00+0000', tz='UTC')
current_date

from src.inference import load_batch_of_features_from_store

features = load_batch_of_features_from_store(current_date)

##NUEVO
##SEGUIR PROBANDO esta guardando el modelo como un array y creo que esta mal
from src.model_registry_api import get_latest_model_from_registry
from src.inference import get_model_predictions

model = get_latest_model_from_registry(model_name='bike_demand_predictor_next_hour', status= 'Production')
predictions = get_model_predictions(model, features)

# from src.inference import (
#     load_model_from_registry,
#     get_model_predictions
# )

# model = load_model_from_registry()
# predictions = get_model_predictions(model, features)

predictions['pickup_hour'] = current_date
predictions

### Save these predictions in the feature store, so they can be later consumed by our Streamlit app

from src.feature_store_api import get_feature_store
import src.config as config

# connect to the feature group
feature_group = get_feature_store().get_or_create_feature_group(
    name=config.FEATURE_GROUP_MODEL_PREDICTIONS,
    version=1,
    description="Predictions generate by our production model",
    primary_key = ['pickup_location_id', 'pickup_hour'],
    event_time='pickup_hour',
)

feature_group.insert(predictions, write_options={"wait_for_job": False})
